Move play function into usecase
