+++ pkg/controller/cli/config/policy.go
@@ -3,6 +3,7 @@ package config
 import (
 	"log/slog"
 
+	"github.com/secmon-lab/alertchain/pkg/chain/core"
 	"github.com/secmon-lab/alertchain/pkg/infra/policy"
 	"github.com/secmon-lab/alertchain/pkg/logging"
 	"github.com/urfave/cli/v3"
@@ -47,4 +48,27 @@ func (x *Policy) Load(pkgName string) (*policy.Client, error) {
 		slog.String("path", x.path),
 	)
 	return policy.New(policy.WithDir(x.path), policy.WithPackage(pkgName))
+}
+
+func (x *Policy) CoreOption() ([]core.Option, error) {
+	var options []core.Option
+
+	if x.Print() {
+		logging.Default().Info("enable print mode")
+		options = append(options, core.WithEnablePrint())
+	}
+
+	alertPolicy, err := x.Load("alert")
+	if err != nil {
+		return nil, err
+	}
+	options = append(options, core.WithPolicyAlert(alertPolicy))
+
+	actionPolicy, err := x.Load("action")
+	if err != nil {
+		return nil, err
+	}
+	options = append(options, core.WithPolicyAction(actionPolicy))
+
+	return options, nil
 }
